+++ src/sll/api/string_api.c
@@ -17,10 +17,6 @@
 
 
 
-static void _object_to_string(const sll_object_t* a,sll_string_t* o);
-
-
-
 static void _write_int(uint64_t v,sll_string_t* o){
 	uint8_t i=0;
 	sll_char_t bf[20];
@@ -36,6 +32,7 @@ static void _write_int(uint64_t v,sll_string_t* o){
 		o->l++;
 	}
 }
+
 
 
 static void _write_char(sll_char_t c,sll_string_t* o){
@@ -92,7 +89,7 @@ static void _write_char(sll_char_t c,sll_string_t* o){
 
 
 
-static void _object_to_string(const sll_object_t* a,sll_string_t* o){
+static void _object_to_string(sll_object_t* a,sll_string_t* o){
 	if (!a->rc){
 		sll_string_increase(o,17);
 		sll_copy_string(SLL_CHAR("<released object>"),o->v+o->l);
@@ -190,10 +187,22 @@ _print_unknown:
 				if (!sll_current_runtime_data||SLL_OBJECT_GET_TYPE(a)>sll_current_runtime_data->tt->l+SLL_MAX_OBJECT_TYPE){
 					goto _print_unknown;
 				}
+				const sll_object_type_data_t* dt=*(sll_current_runtime_data->tt->dt+SLL_OBJECT_GET_TYPE(a)-SLL_MAX_OBJECT_TYPE-1);
+				if (dt->fn.str){
+					_push_call_stack(SLL_CHAR("@sll_api_string_convert"),SLL_MAX_STACK_OFFSET);
+					sll_object_t* v=sll_execute_function(dt->fn.str,&a,1);
+					_pop_call_stack();
+					sll_object_t* str=sll_operator_cast(v,sll_static_int[SLL_OBJECT_TYPE_STRING]);
+					SLL_RELEASE(v);
+					sll_string_increase(o,str->dt.s.l);
+					sll_copy_data(str->dt.s.v,str->dt.s.l,o->v+o->l);
+					o->l+=str->dt.s.l;
+					SLL_RELEASE(str);
+					return;
+				}
 				sll_string_increase(o,3);
 				sll_copy_string(SLL_CHAR("<&:"),o->v+o->l);
 				o->l+=3;
-				const sll_object_type_data_t* dt=*(sll_current_runtime_data->tt->dt+SLL_OBJECT_GET_TYPE(a)-SLL_MAX_OBJECT_TYPE-1);
 				if (!dt->nm.l){
 					_write_int(SLL_OBJECT_GET_TYPE(a),o);
 				}