+++ src/sll/include/sll/api/int.h
@@ -26,6 +26,18 @@ __API_FUNC_DECL(int_to_bin);
 
 /**
  * \flags api compilation_call func optimizable
+ * \name sll_api_int_to_dec
+ * \group int-api
+ * \desc Docs!
+ * \arg I
+ * \ret S
+ */
+__API_FUNC_DECL(int_to_dec)
+
+
+
+/**
+ * \flags api compilation_call func optimizable
  * \name sll_api_int_to_hex
  * \group int-api
  * \desc Docs!