+++ src/sll/io/write.c
@@ -129,10 +129,6 @@ static const sll_node_t* _write_object(sll_file_t* wf,const sll_node_t* o){
 				}
 				return o;
 			}
-		case SLL_NODE_TYPE_DECL_COPY:
-			_write_integer(wf,o->dt.dc.t);
-			_write_integer(wf,o->dt.dc.nm+1);
-			return o+1;
 	}
 	_write_integer(wf,o->dt.ac);
 	sll_arg_count_t l=o->dt.ac;
@@ -245,15 +241,6 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 	for (sll_string_index_t i=0;i<a_dt->st.l;i++){
 		_write_string(a_dt->st.dt+i,wf);
 	}
-	_write_integer(wf,a_dt->ot_it.l);
-	for (sll_object_type_t i=0;i<a_dt->ot_it.l;i++){
-		const sll_object_type_initializer_t* oi=*(a_dt->ot_it.dt+i);
-		_write_integer(wf,oi->l);
-		for (sll_arg_count_t j=0;j<oi->l;j++){
-			_write_integer(wf,(SLL_OBJECT_GET_TYPE_MASK(oi->dt[j].t)<<1)|(!!(oi->dt[j].t&SLL_OBJECT_FLAG_CONSTANT)));
-			_write_integer(wf,oi->dt[j].f);
-		}
-	}
 	_write_integer(wf,a_dt->dbg.l);
 	for (sll_instruction_index_t i=0;i<a_dt->dbg.l;i++){
 		_write_integer(wf,(a_dt->dbg.dt+i)->ii);
@@ -353,7 +340,6 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 				_write_integer(wf,ai->dt.va.l);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_CAST_TYPE:
-			case SLL_ASSEMBLY_INSTRUCTION_TYPE_DECL_COPY:
 				sll_file_write_char(wf,ai->dt.t);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_CALL:
@@ -414,15 +400,6 @@ __SLL_EXTERNAL void sll_write_compiled_node(sll_file_t* wf,const sll_compilation
 	_write_integer(wf,c_dt->st.l);
 	for (sll_string_index_t i=0;i<c_dt->st.l;i++){
 		_write_string(c_dt->st.dt+i,wf);
-	}
-	_write_integer(wf,c_dt->ot_it.l);
-	for (sll_object_type_t i=0;i<c_dt->ot_it.l;i++){
-		const sll_object_type_initializer_t* oi=*(c_dt->ot_it.dt+i);
-		_write_integer(wf,oi->l);
-		for (sll_arg_count_t j=0;j<oi->l;j++){
-			_write_integer(wf,(SLL_OBJECT_GET_TYPE_MASK(oi->dt[j].t)<<1)|(!!(oi->dt[j].t&SLL_OBJECT_FLAG_CONSTANT)));
-			_write_integer(wf,oi->dt[j].f);
-		}
 	}
 	_write_integer(wf,c_dt->fpt.l);
 	for (sll_string_index_t i=0;i<c_dt->fpt.l;i++){