+++ src/sll/print.c
@@ -496,17 +496,6 @@ static const sll_node_t* _print_node_internal(const sll_compilation_data_t* c_dt
 				sll_file_write_char(wf,')');
 				return o;
 			}
-		case SLL_NODE_TYPE_DECL_COPY:
-			PRINT_STATIC_STRING("&:",wf);
-			if (o->dt.dc.nm!=SLL_MAX_STRING_INDEX){
-				PRINT_STATIC_STRING("|#",wf);
-				sll_file_write(wf,(c_dt->st.dt+o->dt.dc.nm)->v,(c_dt->st.dt+o->dt.dc.nm)->l);
-				PRINT_STATIC_STRING("#|",wf);
-			}
-			PRINT_STATIC_STRING(" @",wf);
-			_print_int(o->dt.dc.t,wf);
-			sll_file_write_char(wf,')');
-			return o+1;
 		case SLL_NODE_TYPE_NEW:
 			sll_file_write_char(wf,'.');
 			break;
@@ -1039,13 +1028,6 @@ __SLL_EXTERNAL void sll_print_assembly(const sll_assembly_data_t* a_dt,sll_file_
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_DECL_ZERO:
 				PRINT_STATIC_STRING("DECL 0",wf);
-				if (SLL_ASSEMBLY_INSTRUCTION_FLAG_IS_ANONYMOUS(ai)){
-					PRINT_STATIC_STRING(" (anonymous)",wf);
-				}
-				break;
-			case SLL_ASSEMBLY_INSTRUCTION_TYPE_DECL_COPY:
-				PRINT_STATIC_STRING("DECL @",wf);
-				_print_int(ai->dt.t,wf);
 				if (SLL_ASSEMBLY_INSTRUCTION_FLAG_IS_ANONYMOUS(ai)){
 					PRINT_STATIC_STRING(" (anonymous)",wf);
 				}